Clearly separate generic sync from log_sync
