+++ p2panda-net/src/sync/log_sync/mod.rs
@@ -1,10 +1,11 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
-mod actors;
 mod api;
 mod builder;
 #[cfg(test)]
 mod tests;
 
-pub use api::{LogSync, LogSyncError, LogSyncHandle, LogSyncSubscription};
+pub use api::{LogSync, LogSyncError};
 pub use builder::Builder;
+
+const LOG_SYNC_PROTOCOL_ID: &[u8] = b"p2panda/log_sync/v1";
