+++ p2panda-net/src/sync/mod.rs
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+mod actors;
+mod handle;
+mod log_sync;
+#[cfg(test)]
+mod tests;
+
+pub use handle::{SyncHandle, SyncHandleError, SyncSubscription};
+pub use log_sync::{Builder, LogSync, LogSyncError};
